lab4_for_bd/app/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
black==23.11.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
rpds-py==0.13.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
waitress==2.1.2
Werkzeug==3.0.1
